+++ src/lib.rs
@@ -27,7 +27,7 @@
 //! * [x] Write
 //! * [x] Reg Write
 //! * [x] Action
-//! * [ ] Factory Reset
+//! * [x] Factory Reset
 //! * [x] Reboot
 //! * [ ] Clear
 //! * [ ] Sync Read