+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.1.3:
   * Implement reg_write and action instruction.
+  * Implement factory_reset instruction.
   * Implement custom raw instructions.
   * Include correct license file.